feat(countries): allow decimal kurs
Exchange rates aren't always whole rupiah (e.g., 14.5824 IDR/KRW), so
move `countries.kurs` and `products.kurs` from INTEGER to NUMERIC(12, 4)
and let the form accept decimals. products.kurs has to move with
countries.kurs because product rows snapshot the rate at creation time.

- Migration 025 alters both columns in place; idempotent (guards on the
  current data type so re-running is a no-op).
- schema.sql updated so fresh installs match.
- getCountries / getProducts now coerce r.kurs with Number(), since
  postgres-js returns NUMERIC as a string.
- Kurs input gets step="any" so the browser stops rejecting decimals.

+++ supabase/schema.sql
@@ -5,7 +5,7 @@ CREATE TABLE countries (
   id            SERIAL PRIMARY KEY,
   name          TEXT NOT NULL UNIQUE,
   currency      TEXT NOT NULL DEFAULT '',
-  kurs          INTEGER NOT NULL DEFAULT 0,
+  kurs          NUMERIC(12,4) NOT NULL DEFAULT 0,
   cargo_per_kg  INTEGER NOT NULL DEFAULT 0,
   created_at    TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   updated_at    TIMESTAMPTZ
@@ -44,7 +44,7 @@ CREATE TABLE products (
   gram            INTEGER NOT NULL DEFAULT 0,
   country_id      INTEGER REFERENCES countries(id) ON DELETE RESTRICT,
   valas           NUMERIC(12,2) NOT NULL DEFAULT 0,
-  kurs            INTEGER NOT NULL DEFAULT 0,
+  kurs            NUMERIC(12,4) NOT NULL DEFAULT 0,
   cargo_per_kg    INTEGER NOT NULL DEFAULT 0,
   profit_pct      INTEGER NOT NULL DEFAULT 0,
   operational_fee INTEGER NOT NULL DEFAULT 5000,
